requirements.txt
beautifulsoup4==4.13.3
blinker==1.9.0
boto3==1.36.21
botocore==1.36.21
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
MarkupSafe==3.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.11.2
six==1.17.0
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
openai==1.63.0
redis==5.2.1

